Suppress rendering of top level view only on initial page load. Suppress rendering for the entire view hierarchy.
The fact that we do a root level decision on whether or not to set off the rendering chain has the important implication that functionality render() MUST NOT rely on any other state than the one contained on the view object itself. All other render like functionality that depends on state outside of the view (e. g. cookie, a JSONP request, Bones.user, etc.) needs to be executed on a different stage. Which stage this is to be remains to be determined. This could be a pattern that follows what we've been doing for attach() or ready() on other projects.